Dudik/skip dynamic config test net8 (#7608)
## Summary of changes
Tests that use ClrMD to inspect the live heap for existing objects often
hang on specific .NET versions and architectures. We recently saw this
on .NET 8.0 (both x86 and x64), so we are skipping it for now. If we see
this happening more often, we will need to address it in a different
way.