[Profiler] Add Index to samples and build index vector for libdatadog (#7559)
## Summary of changes

Add index to samples 

## Reason for change
Libdatadog will update the API to create and send .pprof. In that
context, it is needed to map offsets in values vector to specific
profiles used to create .pprof or OTel format.

## Implementation details
Propagate the index from the providers to the samples down to the
profile exporter.

## Test coverage
Add SampleTypeProvider tests

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->