Fix for sending duplicate logs when using Agentless Logging in Azure Function host (#7383)
## Summary of changes

Disables agentless logging (for the Azure Functions Host Process) if we
detect that we are instrumenting an Azure Functions host process.
Previously this would cause us to duplicate every log sent from the
worker process.

This bug fix can be reverted by setting the new configuration key
`DD_LOGS_DIRECT_SUBMISSION_AZURE_FUNCTIONS_HOST_ENABLED` to `true`.

## Reason for change

If customers are running with the Isolated Azure Function model we will
be instrumenting two applications, one is the function host process the
other is the worker process. If they have direct log submission enabled
the function host ends up duplicating the logs from the function process
which results in us shipping two nearly identical logs.

This behavior isn't ideal as the duplicate log is likely not valuable so
we've disabled agentless logging in the azure functions host process.

## Implementation details

Added `IsRunningInAzureFunctionsHost()` to `EnvironmentHelpers.cs` which
allows for a rough detection of whether we are running on the function
host using the following logic:

- Is `FUNCTIONS_WORKER_RUNTIME` present AND set to `dotnet-isolated`?
- Are both `--functions-worker-id` or `--workerId` NOT present in the
command line?

If both are true we treat that scenario as being running in the function
host - otherwise we are likely the worker process

I wasn't able to find a more robust way of checking, but when looking at
the various log output that I had the `--functions-worker-id` or
`--workerId` seemed to always be called by the function host.

```
[2025-10-03T16:01:42.901Z] Reading functions metadata (Worker)
[2025-10-03T16:01:47.176Z] {
[2025-10-03T16:01:47.177Z]   "ProcessId": 71080,
[2025-10-03T16:01:47.178Z]   "RuntimeIdentifier": "win-x64",
[2025-10-03T16:01:47.179Z]   "WorkerVersion": "2.0.0.0",
[2025-10-03T16:01:47.180Z]   "ProductVersion": "2.0.0\u002Bd8b5fe998a8c92819b8ee41d2569d2525413e9c5",
[2025-10-03T16:01:47.181Z]   "FrameworkDescription": ".NET 9.0.9",
[2025-10-03T16:01:47.182Z]   "OSDescription": "Microsoft Windows 10.0.26100",
[2025-10-03T16:01:47.183Z]   "OSArchitecture": "X64",
[2025-10-03T16:01:47.184Z]   "CommandLine": "C:\\Users\\steven.bouwkamp\\source\\repos\\dd-trace-dotnet\\artifacts\\bin\\Samples.AzureFunctions.V4Isolated.AspNetCore\\debug_net9.0\\Samples.AzureFunctions.V4Isolated.AspNetCore.dll --host 127.0.0.1 --port 65401 --workerId e94d23fd-cd3c-4780-a3e3-4980d7b0f644 --requestId 6dba68ac-1954-466a-aeb4-9570cc9b12c2 --grpcMaxMessageLength 2147483647 --functions-uri http://127.0.0.1:65401/ --functions-worker-id e94d23fd-cd3c-4780-a3e3-4980d7b0f644 --functions-request-id 6dba68ac-1954-466a-aeb4-9570cc9b12c2 --functions-grpc-max-message-length 2147483647"
[2025-10-03T16:01:47.185Z] }
```

- Added `DD_LOGS_DIRECT_SUBMISSION_AZURE_FUNCTIONS_HOST_ENABLED` that
defaults to `false` to disable the duplicate logs from being sent.

## Test coverage

- Added a new test project and test for having host logs enabled /
disabled.
- The reason why I added a new test project instead of re-using an
existing one was because when I re-ran the Function application in our
tests multiple times the `func.exe` would fail to obtain a lock and
would need to wait some period of time to for recovery after each
subsequent run of the same application. I think this is because we end
the `func.exe` process with a `Kill()`. Making a new project wasn't
ideal but was a quick and simple workaround.

## Other details
Fixes SLES-2364

<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->