Revert "[SINT-4157] sanitise ultimate-pipeline.yml" (#7627)
Reverts DataDog/dd-trace-dotnet#7582

This is preventing us to run the action as expected as the regex appears
to be incorrect.
Since we must trigger this ourselves there really isn't a risk here at
the moment, will let the SINT team follow up to resolve.

This is blocking the testing of
https://github.com/DataDog/dd-trace-dotnet/pull/7514


https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=188595&view=logs&j=8299a7f3-61e2-5626-a12e-85b7b20deab1&t=e26f8d61-df24-5d93-564c-322f7221200d