[Dynamic Instrumentation] Ensure DI metric probe does not send automatic telemetry (#7621)
## Summary of changes
Ensure that we do not send automatic telemetry with the DogStatsd metric
probe.

## Reason for change
Since version
[2.38.0](https://github.com/DataDog/dd-trace-dotnet/releases/tag/v2.38.0),
DogStatsd is created without telemetry, which is also beneficial for DI.
([PR](https://github.com/DataDog/dd-trace-dotnet/pull/4580))
This PR ensures that the behavior from versions prior to 2.38.0 will not
return in the future if the DogStatsd creation logic changes.