Add successful vendoring of Google.Protobuf.
Note: This does not replace the base namespace of the vendored Google.Protobuf code so that we can import the Grpc.Tools NuGet package at build time to generate the C# Protobuf classes. In a separate commit we can try to rename the namespace and generate the C# Protobuf files ahead-of-time.