Clone the prod dd-trace-dotnet microbenchmark branch for microbenchmarks (#7646)
## Summary of changes

Update the Benchmarking Platform branch for microbenchmarks to be
`dd-trace-dotnet/micro`.

## Reason for change

The previous branch, `augusto/dd-trace-dotnet-micro-bp-ui` was being
used to 1. test uploading to the BP UI and to 2. facilitate merging
changes to `dd-trace-dotnet/micro` without having benchmarks on
dd-trace-dotnet fail during this merge[^1].

https://datadoghq.atlassian.net/browse/APMSP-2287

## Implementation details

## Test coverage

Test run on the CI:
https://gitlab.ddbuild.io/DataDog/apm-reliability/dd-trace-dotnet/-/jobs/1176018864

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->

[^1]: To merge breaking changes on `dd-trace-dotnet` and on
`benchmarking-platform @ dd-trace-dotnet/micro`, one needs to first
merge changes on the `benchmarking-platform`, and merge changes on
`dd-trace-dotnet` immediately after. If merging on `dd-trace-dotnet`
took a bit of time (on the order of ~1h), new commits on
`dd-trace-dotnet` within that period of time would have failing
benchmarks.