Minor refactor and added tests for `DynamicConfigurationManager` (#7652)
## Summary of changes

- Adds some unit tests for some of the `APM_TRACING` logic in
`DynamicConfigurationManager`
- ~Reduce duration of `lock`~ Remove `lock`
- Minor reduction in allocations

## Reason for change

Rebasing #7525 after #7536 flagged that there were some conflicts. In
working out how to resolve them, noted some possibilities for minor
improvements

## Implementation details

- Extracted existing functionality into a testable method
- Added unit tests for the existing behaviour
- Tweaks

Technically I believe this is _always_ single-threaded, so we _should_
be able to remove the `lock` and also reduce some `ToList()` etc, but we
don't _explicitly_ enforce that, so I'm not sure whether to do it or not
tbh 🤔 Thoughts? Given [this
comment](https://github.com/apps/chatgpt-codex-connector) I'm inclined
to remove it entirely. EDIT: we removed the lock entirely in the end for
simplicity

## Test coverage

More now

## Other details

Reviewing commit-by-commit shows the changes more clearly 🙂