[OTEL] Switches OTLP Tests To dd-apm-test-agent Instead of MockTracerAgent (#7656)
## Summary of changes
Making use of the
[dd-apm-test-agent](https://github.com/DataDog/dd-apm-test-agent)
container to enable OTLP metrics support for GRPC and HTTP, then gets
then as JSON the container for snapshot validation.

## Reason for change
To easily and more reliably test GRPC and ongoing OTLP support and
validation in the .NET tracer without needing to build GRPC servers or
change our agent for each new signal we plan to support.

## Implementation details
Added the container to run for the integrations tests in the
corresponding host name making use of either port `4317` or 4318`:

1. Pings `test/session/clear` to clear the session at the start of the
tests
2. Pings `/test/session/metrics` endpoint for the generated metrics

## Test coverage
Updated the snapshots format but the same 4 should pass even for GRPC
now that added to run for OTEL.

## Other details
<!-- Fixes #{issue} -->


<!--  ⚠️ Note:

Where possible, please obtain 2 approvals prior to merging. Unless
CODEOWNERS specifies otherwise, for external teams it is typically best
to have one review from a team member, and one review from apm-dotnet.
Trivial changes do not require 2 reviews.

MergeQueue is NOT enabled in this repository. If you have write access
to the repo, the PR has 1-2 approvals (see above), and all of the
required checks have passed, you can use the Squash and Merge button to
merge the PR. If you don't have write access, or you need help, reach
out in the #apm-dotnet channel in Slack.
-->