Add fix for whitespace in exporter settings and metrics URL (#7747)
## Summary of changes

- Trim whitespace in the exporter settings values before trying to use
them
- Fix a bug where having a whitespace in the `DD_DOGSTATSD_SOCKET`
variable would cause incorrect (and invalid) UDS config

## Reason for change

This manifested in a different test in a branch, where an empty string
passed in `DD_DOGSTATSD_SOCKET` was causing us to use UDS for metrics
even through it's not available and not _really_ set.

## Implementation details

- `Trim()` the string variables (I don't think there's a good reason not
to?)
- As we're trimming, we can switch to `IsNullOrEmpty` instead of
`IsWhitespace`
- Treat empty `DD_DOGSTATSD_SOCKET` as not set

Alternatively, we could not trim, stick to using `IsWhitespace`, and
just update the check I fixed to use `IsWhitespace` 🤷‍♂️

## Test coverage

Added a couple of unit tests. Only one test is for this specific issue,
the others were just part of my investigation, and seemed reasonable.

## Other details

Discovered as part of
https://github.com/DataDog/dd-trace-dotnet/pull/7724 and
https://datadoghq.atlassian.net/browse/LANGPLAT-819