Fix Asm benchmark and stop-reinitializing tracer (#7750)
## Summary of changes

- Fix ASM benchmarks by not returning `null` from
`EmptyDatadogTracer.PerTraceSettings`
- Stop re-initializing `TracerSettings` for every execution of the ASM
benchmark

## Reason for change

The ASM benchmarks have been broken since
https://github.com/DataDog/dd-trace-dotnet/pull/7543 (I think). This is
because the `EmptyDatadogTracer` stub used in the benchmark returns
`null` from `PerTraceSettings` (which can't happen in practice).

Additionally, noticed that the benchmark is repeatedly creating a new
`TracerSettings` object with every execution, which will add noise and
be much more expensive than real life.

## Implementation details

- Ensure `EmptyDatadogTracer.PerTraceSettings` returns a "real" value
- Stop rebuilding `TracerSettings` with every execution

## Test coverage

This is the test