Add partial trust/legacy CAS detection to dd-trace and dd-dotnet (#7779)
## Summary of changes

- Fixes a bug in AppSettings extraction in dd-dotnet
- Adds additional checks to `iis check <site>` to check for partial
trust or the legacy CAS mode

## Reason for change

We recently [had an
escalation](https://datadoghq.atlassian.net/browse/APMS-17451) where
legacy CAS mode was the problem, but it took a long time to identify.
Adding a check to the tool should improve the time to resolution (and
allow customers to identify the issue themselves) more quickly.

## Implementation details

Read the `system.web/trust` node from the application's web.config and
check the `level` and `legacyCasModel` properties. If they reveal a
problem, report it. I've largely ignored the .NET Core/.NET Framework
aspect here, as it's super niche. Worst case we report something that
doesn't actually matter (but it would be a _weird_ config setup).

Also noticed that `GetAppSettings` wasn't getting the correct settings
for the app,.

## Test coverage

Manually tested it all locally with various configurations, setting the
values in root and leaf configs, etc and confirmed it all works as
expected (config flattening including on nested apps).

## Other details

Theoretically `runtime/NetFx40_LegacySecurityPolicy` should play into it
too, but AFAICT, that doesn't apply to aspnet apps, so it's not the
focus of this PR. If we need to, we could check it for non-iis apps, but
that's pretty niche right now.