Add fix for persistent `Process.Start` version-conflict issue (#7789)
## Summary of changes

- Adds a workaround for the version-conflict issue that occurs on app
startup on Linux
- Allow doing call target modification of version conflict dll
- Fix a bug in CallTarget instrumentation where we get the assembly
reference wrong

## Reason for change

As part of app startup, on _some_ linux distros, we shell out to `stat`
to build the container tags/entity ID. We don't want to trace this
`Process.Start()` call, so in
https://github.com/DataDog/dd-trace-dotnet/pull/5280 we added a flag to
skip instrumenting these calls. However, this fix relies on a
`[ThreadStatic]` variable, and in version-conflict scenarios (2.x.x
manual, 3.x.x automatic) we end up still instrumenting this call, which
causes recursion in `Tracer` initialization and
[errors](https://app.datadoghq.com/error-tracking?query=service%3Ainstrumentation-telemetry-data%20source%3Adotnet%20%40tracer_version%3A3.29.0.0%20-%40error.is_crash%3Atrue&et-side=activity&order=total_count&refresh_mode=sliding&source=all&sp=%5B%7B%22p%22%3A%7B%22issueId%22%3A%22cbce5fc2-3adf-11f0-a4be-da7ad0900002%22%7D%2C%22i%22%3A%22error-tracking-issue%22%7D%5D&view=spans&from_ts=1761134402500&to_ts=1762344002500&live=true).

> Note that since https://github.com/DataDog/dd-trace-dotnet/pull/7453
we don't do a process start at all, but that doesn't help in this
situation, because it's the 2.x.x library that's doing the
`Process.Start()`

## Implementation details

- Use standard call target instrumentation on the 2.x.x version of
`Datadog.Trace` (i.e. version conflict only)
- Hook the `ProcessHelpers.StartWithDoNotTrace()` method, and set the
3.x.x `_doNotTrace` variable for the duration of the method call
- Tweak the Rejit handler so that we _do_ rejit/call target the
Datadog.Trace 2.x.x module (but not the 3.x.x module)
- Fix a bug in the module builder which was incorrectly injecting a
reference to the 2.x.x assembly instead of the 3.x.x assembly

## Test coverage

We were already working around this issue in our VersionConflict tests,
so I removed the workaround, confirmed that the test failed, then made
the fix, and confirmed the tests pass again.

## Other details

This will only help for customers using a manual version of 2.49.0+
(when we introduced the `StartWithDoNotTrace()` call). I think that's
good enough support.

---------

Co-authored-by: Lucas Pimentel <lucas.pimentel@datadoghq.com>