Fix timestamp reporting in OTLP Logs to accurately represent UTC time.
Previously, we had started with a UTC DateTime for the LogRecord object and when encoding it as a Unix epoch time we applied a second "local->UTC time" via using DateTime.ToUniversalTime(). This resulted in OTLP logs from the PST being reported as having a local time 7 hours in the future.