[Logs] Fix OTLP Logs timestamp during serialization (#7814)
## Summary of changes
Fixes the serialization of the OTLP Logs timestamps fields.

## Reason for change
Before this change, we would start with a UTC DateTime for the LogRecord
object and then when we would encode it as a Unix epoch time we applied
a "local->UTC time" conversion by using DateTime.ToUniversalTime(),
which shifted the time. This resulted in OTLP logs from the PST timezone
being reported as having a local time 7 hours in the future.

## Implementation details
- Change the `Timestamp` property of
`Datadog.Trace.OpenTelemetry.Logs.LogPoint` to have type
`DateTimeOffset` instead of `DateTime` so we're always dealing with a
UTC time, and initialize the value to `DateTimeOffset.UtcNow`
- Remove the previous encoding of DateTime->Unix nanoseconds and replace
it with a call to the existing extension method
`TimeExtensions.ToUnixTimeNanoseconds`, which is already used for
serializing timestamps for Metrics and Traces

## Test coverage
Add a regression test in `OpenTelemetrySdkTests.SubmitsOtlpLogs` by
asserting that the timestamp in the OTLP log payload is in the expected
range (application lifetime) before scrubbing the data and running the
snapshot test.

## Other details
N/A