Repurpose `PublicApiAnalyzer` as `InternalForTestingAnalyzer`  (#7788)
## Summary of changes

Repurposes the now-defunct `PublicApiAnalyzer` to an
`InternalForTestingAnalyzer`

## Reason for change

We have various methods in `Datadog.Trace` that should never be called
by production code, and are only exposed to make testing functionality
easier. Currently we rely on method names and comments to avoid this.
With this change, we can add an attribute instead, and using the method
from Datadog.Trace etc becomes a build error.

## Implementation details

- Basically just renamed `PublicApi*` -> `InternalForTesting*`
- Decorated methods that were currently relying on comments/naming to
avoid being called

## Test coverage

Refactored the analyzer tests, so should still be covered the same

## Other details

Note that there are a bunch of comments for methods `// internal for
testing` which are a bit different. In those cases, we _do_ call it in
Datadog.Trace, the comment is just explaining why the method isn't
`private`. Those remain as-is, this attribute is purely if you have a
"don't call this method in production code, use that other one instead"
method 😄

Part of a stack

- https://github.com/DataDog/dd-trace-dotnet/pull/7786
- https://github.com/DataDog/dd-trace-dotnet/pull/7787
- https://github.com/DataDog/dd-trace-dotnet/pull/7788 👈