Add support for `MongoDB.Driver` v3.5+ (#7808)
## Summary of changes

Add support for `MongoDB.Driver` v3.5+

## Reason for change

We didn't support 3.5. immediately, as it required using a new version
of mongodb in CI - the existing version wouldn't work. We were running
into capacity reasons, so we figured we would punt on testing it till
later when we could update the VMs.

_Unfortunately_, it turns out 3.5 introduced a bunch of internal
changes, which means instrumentation is broken for 3.5 🤦‍♂️ This adds
support for it back.

## Implementation details

- Bump the version of mongo tested (this is already pulled on the VMs)
- Fix the integration

## Test coverage

Covered by existing.

I pinned testing of the latest 3.4.x too, but ditched some of the
earlier 2.x. versions. We were testing 4 different 2.x.x versions, which
seemed a bit overkill, seeing as they all hit the same code paths, and
haven't changed in years.

## Other details

https://datadoghq.atlassian.net/browse/APMS-17799